added feature login with cookie

+++ src/Components/LoginComponent.js
@@ -12,8 +12,9 @@ export default class LoginComponent extends React.Component {
   constructor (props) {
     super(props)
     this.state = {
-      no_rek: '',
+      no_rek: null,
       loggedIn:false,
+      error:false,
     }
     this.onChange = this.onChange.bind(this)
     this.onLogin = this.onLogin.bind(this)
@@ -23,19 +24,43 @@ export default class LoginComponent extends React.Component {
     const re = /^[0-9\b]+$/
     if (re.test(event.target.value)) {
       this.setState({ no_rek: event.target.value })
-      console.log(this.state.no_rek);
+    }
+  }
+
+  onCallback(rekeningExist){
+    if(rekeningExist){
+      console.log("wakgeng")
+      console.log(this.state.no_rek)
+      localStorage.setItem("cookieBankPro", this.state.no_rek)
+      this.setState({loggedIn:true});
+      this.props.renderNavbar();
+    } else {
+      this.setState({error:true});
     }
   }
 
   onLogin(e){
-    localStorage.setItem("cookieBankPro", this.state.no_rek)
-    this.setState({loggedIn:true});
-    this.props.renderNavbar();
+    console.log(this.state.no_rek);
+    this.validateRekening(this.state.no_rek, this.onCallback.bind(this));
     e.preventDefault();
   }
 
+  validateRekening(AccNo, cb){
+    let res = false;
+    var soap = require('soap');
+    var url = 'http://3.93.238.160:8080/bankprowebservice-1.0-SNAPSHOT/NewWebService?wsdl';
+    var args = {Rekening: AccNo};
+    soap.createClient(url, function(err, client) {
+      client.validateRekening(args, function(err, result) {
+        console.log(result['return']);
+        res = result['return'];
+        cb(res)
+      });
+    });
+  }
+
   render () {
-    const { error } = this.state;
+    const { error } = this.state.error;
     return (
       <Container className="center-container">
         <div className="card">
@@ -52,10 +77,7 @@ export default class LoginComponent extends React.Component {
                   placeholder="your account number"
                   onChange={(e) => this.onChange(e)}
                 />
-                {error && <div color='red'
-                                   error={error}
-                                   content="There's no such account. Try again!"
-                />}
+                {error && <div>There's no such account. Try again!</div>}
               </FormGroup>
             </Col>
             <Button type="submit" style={{ backgroundColor: '#21242D' }} >Login</Button>
